Moves accuracy metric functions to new module file
