+++ src/flownet/ahm/__init__.py
@@ -2,6 +2,8 @@ from ._assisted_history_matching import (
     AssistedHistoryMatching,
     delete_simulation_output,
     save_iteration_parameters,
+)
+from ._ahm_iteration_analytics import (
     save_iteration_analytics,
 )
 from ._run_ahm import run_flownet_history_matching
